# use EndpointBuilder in CDS