--- a/pilot/pkg/model/push_context.go
+++ b/pilot/pkg/model/push_context.go
@@ -1197,7 +1197,7 @@ func (ps *PushContext) getExportedDestinationRuleFromNamespace(owningNamespace s
 // IsClusterLocal indicates whether the endpoints for the service should only be accessible to clients
 // within the cluster.
 func (ps *PushContext) IsClusterLocal(service *Service) bool {
-	if service == nil {
+	if ps == nil || service == nil {
 		return false
 	}
 	return ps.clusterLocalHosts.IsClusterLocal(service.Hostname)