--- a/pilot/pkg/networking/core/v1alpha3/cluster_cache.go
+++ b/pilot/pkg/networking/core/v1alpha3/cluster_cache.go
@@ -22,6 +22,7 @@ import (
 
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/networking/util"
+	"istio.io/istio/pilot/pkg/xds/endpoints"
 	"istio.io/istio/pkg/config/schema/kind"
 	"istio.io/istio/pkg/util/hash"
 )
@@ -37,13 +38,14 @@ type clusterCache struct {
 	clusterName string
 
 	// proxy related cache fields
-	proxyVersion   string         // will be matched by envoyfilter patches
-	locality       *core.Locality // identifies the locality the cluster is generated for
-	proxyClusterID string         // identifies the kubernetes cluster a proxy is in
-	proxySidecar   bool           // identifies if this proxy is a Sidecar
-	hbone          bool
-	proxyView      model.ProxyView
-	metadataCerts  *metadataCerts // metadata certificates of proxy
+	proxyVersion    string         // will be matched by envoyfilter patches
+	locality        *core.Locality // identifies the locality the cluster is generated for
+	proxyClusterID  string         // identifies the kubernetes cluster a proxy is in
+	proxySidecar    bool           // identifies if this proxy is a Sidecar
+	hbone           bool
+	proxyView       model.ProxyView
+	metadataCerts   *metadataCerts // metadata certificates of proxy
+	endpointBuilder endpoints.EndpointBuilder
 
 	// service attributes
 	http2          bool // http2 identifies if the cluster is for an http2 service
@@ -124,6 +126,8 @@ func (t *clusterCache) Key() any {
 	}
 	h.Write(Separator)
 
+	t.endpointBuilder.WriteHash(h)
+
 	return h.Sum64()
 }
 
@@ -142,6 +146,9 @@ func (t *clusterCache) DependentConfigs() []model.ConfigHash {
 		items := strings.Split(efKey, "/")
 		configs = append(configs, model.ConfigKey{Kind: kind.EnvoyFilter, Name: items[1], Namespace: items[0]}.HashCode())
 	}
+
+	// For now, this matches EndpointBuilder's DependentConfigs. No need to duplicate them.
+
 	return configs
 }
 
@@ -166,8 +173,10 @@ func (c cacheStats) merge(other cacheStats) cacheStats {
 }
 
 func buildClusterKey(service *model.Service, port *model.Port, cb *ClusterBuilder, proxy *model.Proxy, efKeys []string) clusterCache {
+	clusterName := model.BuildSubsetKey(model.TrafficDirectionOutbound, "", service.Hostname, port.Port)
+	dr := proxy.SidecarScope.DestinationRule(model.TrafficDirectionOutbound, proxy, service.Hostname)
 	return clusterCache{
-		clusterName:     model.BuildSubsetKey(model.TrafficDirectionOutbound, "", service.Hostname, port.Port),
+		clusterName:     clusterName,
 		proxyVersion:    cb.proxyVersion,
 		locality:        cb.locality,
 		proxyClusterID:  cb.clusterID,
@@ -178,10 +187,17 @@ func buildClusterKey(service *model.Service, port *model.Port, cb *ClusterBuilde
 		downstreamAuto:  cb.sidecarProxy() && port.Protocol.IsUnsupported(),
 		supportsIPv4:    cb.supportsIPv4,
 		service:         service,
-		destinationRule: proxy.SidecarScope.DestinationRule(model.TrafficDirectionOutbound, proxy, service.Hostname),
+		destinationRule: dr,
 		envoyFilterKeys: efKeys,
 		metadataCerts:   cb.metadataCerts,
 		peerAuthVersion: cb.req.Push.AuthnPolicies.GetVersion(),
 		serviceAccounts: cb.req.Push.ServiceAccounts(service.Hostname, service.Attributes.Namespace, port.Port),
+		endpointBuilder: endpoints.NewCDSEndpointBuilder(
+			proxy,
+			cb.req.Push,
+			clusterName,
+			model.TrafficDirectionOutbound, "", service.Hostname, port.Port,
+			service, dr,
+		),
 	}
 }