Changes:
* Use EndpointBuilder in favor of `buildLocalityLbEndpoints` in cluster builder. That cluster builder method is a partial duplication of a lot of the same logic we have for EDS. 
* Enable multi-network at CDS level

Benchmarks are  no longer affected. 

```
                                              │ /home/landow/old.txt │        /home/landow/new.txt        │
                                                │        sec/op        │   sec/op     vs base               │
ClusterGeneration/gateways-24                              4.844m ± 2%   4.875m ± 4%       ~ (p=0.529 n=10)
ClusterGeneration/gateways-shared-24                       4.864m ± 4%   4.873m ± 3%       ~ (p=0.796 n=10)
ClusterGeneration/knative-gateway-24                       12.55µ ± 2%   12.71µ ± 4%       ~ (p=0.393 n=10)
ClusterGeneration/http-24                                  868.8µ ± 4%   877.8µ ± 4%       ~ (p=0.631 n=10)
ClusterGeneration/tcp-24                                   867.0µ ± 4%   874.0µ ± 3%       ~ (p=0.579 n=10)
ClusterGeneration/tls-24                                   881.4µ ± 4%   884.2µ ± 4%       ~ (p=0.912 n=10)
ClusterGeneration/auto-24                                 1021.5µ ± 5%   994.1µ ± 7%       ~ (p=0.247 n=10)
ClusterGeneration/telemetry-api-24                         2.420m ± 5%   2.461m ± 3%       ~ (p=0.436 n=10)
ClusterGeneration/virtualservice-24                        565.5µ ± 3%   573.3µ ± 4%       ~ (p=0.143 n=10)
ClusterGeneration/serviceentry-workloadentry-24            3.318m ± 5%   3.238m ± 6%       ~ (p=0.529 n=10)
geomean                                                    990.7µ        992.7µ       +0.21%

                                                │ /home/landow/old.txt │        /home/landow/new.txt         │
                                                │        kb/msg        │   kb/msg    vs base                 │
ClusterGeneration/gateways-24                               390.3 ± 0%   390.3 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/gateways-shared-24                        390.3 ± 0%   390.3 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/knative-gateway-24                        1.058 ± 0%   1.058 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/http-24                                   86.34 ± 0%   86.34 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/tcp-24                                    86.95 ± 0%   86.95 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/tls-24                                    86.95 ± 0%   86.95 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/auto-24                                   101.8 ± 0%   101.8 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/telemetry-api-24                          172.9 ± 0%   172.9 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/virtualservice-24                         39.70 ± 0%   39.70 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/serviceentry-workloadentry-24             379.1 ± 0%   379.1 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                     88.01        88.01       +0.00%
¹ all samples are equal

                                                │ /home/landow/old.txt │          /home/landow/new.txt          │
                                                │    resources/msg     │ resources/msg  vs base                 │
ClusterGeneration/gateways-24                              1.003k ± 0%     1.003k ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/gateways-shared-24                       1.003k ± 0%     1.003k ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/knative-gateway-24                        3.000 ± 0%      3.000 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/http-24                                   105.0 ± 0%      105.0 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/tcp-24                                    105.0 ± 0%      105.0 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/tls-24                                    105.0 ± 0%      105.0 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/auto-24                                   105.0 ± 0%      105.0 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/telemetry-api-24                          411.0 ± 0%      411.0 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/virtualservice-24                         105.0 ± 0%      105.0 ± 0%       ~ (p=1.000 n=10) ¹
ClusterGeneration/serviceentry-workloadentry-24             411.0 ± 0%      411.0 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                     151.8           151.8       +0.00%
¹ all samples are equal



```


cc @kdorosh @howardjohn