@stevenctl 2023-06-29
Next, I'd like to consolidate on shards vs service instances but that has implications for cds cache.