@istio-testing 2023-06-29
@stevenctl: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
release-notes_istio | 4a7152c8f477912275fc3f12d859eef4906b78c7 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45733/release-notes_istio/1674448195884683264) | true | `/test release-notes`
unit-tests-arm64_istio | 4a7152c8f477912275fc3f12d859eef4906b78c7 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45733/unit-tests-arm64_istio/1674448182756511744) | true | `/test unit-tests-arm64`
lint_istio | 4a7152c8f477912275fc3f12d859eef4906b78c7 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45733/lint_istio/1674448194198573056) | true | `/test lint`
gencheck_istio | 4a7152c8f477912275fc3f12d859eef4906b78c7 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45733/gencheck_istio/1674448195041628160) | true | `/test gencheck`
unit-tests_istio | 4a7152c8f477912275fc3f12d859eef4906b78c7 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45733/unit-tests_istio/1674448182706180096) | true | `/test unit-tests`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->